rename all the RT methods
