+++ packages/core/src/rooms/index.ts
@@ -9,7 +9,6 @@ export interface BaseRoomInterface<
 }
 
 export * from './methods'
-export * from './methodsRT'
 export * from './RoomSessionRecording'
 export * from './RoomSessionPlayback'
 export * from './RoomSessionStream'
